ENG-10259: remove Statement in favor of Queryer interface
