+++ voltdbclient/conn.go
@@ -24,7 +24,11 @@ import (
 	"io"
 	"net"
 	"reflect"
+	"bytes"
+	"sync/atomic"
 )
+
+var qHandle int64 = 0  // each query has a unique handle.
 
 // connectionData are the values returned by a successful login.
 type connectionData struct {
@@ -95,11 +99,35 @@ func OpenConn(connInfo string) (*VoltConn, error) {
 }
 
 func (vc VoltConn) Prepare(query string) (driver.Stmt, error) {
+	panic("Prepare is not supported by a Volt Connection")
+}
+
+func (vc VoltConn) Query(query string, args []driver.Value) (driver.Rows, error) {
 	if !vc.isOpen {
 		return nil, errors.New("Connection is closed")
 	}
-	vs := newVoltStatement(&vc, &vc.writer, vc.netListener, query)
-	return *vs, nil
+	handle := atomic.AddInt64(&qHandle, 1)
+	c := vc.netListener.registerQuery(handle)
+	if err := vc.serializeQuery(&vc.writer, query, handle, args); err != nil {
+		vc.netListener.removeQuery(handle)
+		return VoltRows{}, err
+	}
+	return <-c, nil
+}
+
+func (vc VoltConn) QueryAsync(query string, args []driver.Value) (*VoltQueryResult, error) {
+	if !vc.isOpen {
+		return nil, errors.New("Connection is closed")
+	}
+	handle := atomic.AddInt64(&qHandle, 1)
+	c := vc.netListener.registerQuery(handle)
+	vqr := newVoltQueryResult(&vc, handle, c)
+	vc.registerQuery(handle, vqr)
+	if err := vc.serializeQuery(&vc.writer, query, handle, args); err != nil {
+		vc.netListener.removeQuery(handle)
+		return nil, err
+	}
+	return vqr, nil
 }
 
 func (vc VoltConn) Drain(vqrs []*VoltQueryResult) {
@@ -174,4 +202,81 @@ func (vc VoltConn) registerQuery(handle int64, vcr *VoltQueryResult) {
 
 func (vc VoltConn) removeQuery(han int64) {
 	delete(vc.queries, han)
+}
+
+type VoltQueryResult struct {
+	conn *VoltConn
+	han int64
+	ch <-chan driver.Rows
+	rows driver.Rows
+	err error
+	active bool
+}
+
+func newVoltQueryResult(conn *VoltConn, han int64, ch <-chan driver.Rows) *VoltQueryResult {
+	var vqr = new(VoltQueryResult)
+	vqr.conn = conn
+	vqr.han = han
+	vqr.ch = ch
+	vqr.active = true
+	return vqr
+}
+
+func (vqr *VoltQueryResult) Result() (driver.Rows, error) {
+	if !vqr.active {
+		return vqr.rows, vqr.err
+	} else {
+		rows := <-vqr.ch
+		vqr.setRows(rows)
+		return vqr.rows, nil
+	}
+}
+
+func (vqr *VoltQueryResult) channel() <-chan driver.Rows {
+	return vqr.ch
+}
+
+func (vqr *VoltQueryResult) handle() int64 {
+	return vqr.han
+}
+
+func (vqr *VoltQueryResult) isActive() bool {
+	return vqr.active
+}
+
+func (vqr *VoltQueryResult) setError(err error) {
+	if !vqr.active {
+		panic("Tried to set error on inactive query result")
+	}
+	vqr.err = err
+	vqr.conn.removeQuery(vqr.han)
+	vqr.active = false
+}
+
+func (vqr *VoltQueryResult) setRows(rows driver.Rows) {
+	if !vqr.active {
+		panic("Tried to set rows on inactive query result")
+	}
+	vqr.rows = rows
+	vqr.conn.removeQuery(vqr.han)
+	vqr.active = false
+}
+
+
+func (vc VoltConn) serializeQuery(writer *io.Writer, procedure string, handle int64, args []driver.Value) error {
+
+	var call bytes.Buffer
+	var err error
+
+	// Serialize the procedure call and its params.
+	// Use 0 for handle; it's not necessary in pure sync client.
+	if call, err = serializeStatement(procedure, handle, args); err != nil {
+		return err
+	}
+
+	var netmsg bytes.Buffer
+	writeInt(&netmsg, int32(call.Len()))
+	io.Copy(&netmsg, &call)
+	io.Copy(*writer, &netmsg)
+	return nil
 }
